+++ data/style.css
@@ -1,0 +1,98 @@
+:root {
+    --RUSSIAN_VIOLETT:  #241E4E;
+    --GREEN_COLOR:      #399E5A;
+    --RED_COLOR:        #CE6C47;
+    --GAINSBORO:        #DFDFDF;
+    --TEAL_BLUE:        #40798C;
+ }
+
+ html {
+    font-family: Helvetica;
+    display: inline-block;
+    margin: 0px auto;
+    text-align: center;
+  }
+  h1{
+    color: #767e8d;
+    padding: 2vh;
+  }
+  p{
+    font-size: 1.5rem;
+  }
+  .button {
+    display: inline-block;
+    background-color: #008CBA;
+    border: none;
+    border-radius: 4px;
+    color: white;
+    padding: 16px 40px;
+    text-decoration: none;
+    font-size: 30px;
+    margin: 2px;
+    cursor: pointer;
+  }
+  .button2 {
+    background-color: #f44336;
+  }
+  body {
+    background-color: #41635a;
+    color: var(--GAINSBORO);
+    text-align: center;
+    font-family: Lato, Helvetica, Roboto, sans-serif;
+    border: 1em;
+  }
+  caption {
+    caption-side: top;
+    padding: 10px;
+    font-weight: bold;
+  }
+  table {
+  border-collapse: separate;
+  width:50%;
+  margin-left:auto;
+  margin-right:auto;
+  }
+  th {
+  padding: 12px;
+  background-color: #a7c3f1;
+  color: white;
+  }
+  tr {
+    padding: 12px;
+  }
+  td {
+    width:50%;
+    padding: 12px;
+    text-align: right;
+  }
+  td:last-child {
+    width:50%;
+    background-color: #e0f7fa;
+    color: #00796b;
+    padding: 5px;
+    text-align: left;
+    box-shadow: inset 2px 2px 4px rgba(0, 0, 0, 0.4);
+    text-shadow: 2px 2px 4px #000000;
+}  
+  .sensor {
+  color:white;
+  font-weight: bold;
+  background-color: #bcbcbc;
+  padding: 1px;
+  }
+  /* Box styles */
+  .myBox {
+  border: 2px solid;
+  border-radius: 5px;
+  margin: auto;
+  padding: 5px;
+  font: 24px/36px sans-serif;
+  text-align: left;
+  width: auto;
+  height: auto;
+  overflow-y: scroll;
+  }
+  .valuecolumn {
+    background-color: #bcbcbc;
+  }
+
