+++ platformio.ini
@@ -13,12 +13,8 @@ platform = espressif32
 board = az-delivery-devkit-v4
 framework = arduino
 monitor_speed = 115200
-monitor_port = COM7
-;build_type = debug
-;check_skip_packages = yes
 build_flags = 
 	-DBUILD_PROCESS
-	;!powershell -NonInteractive -NoLogo -NoProfile -File .\GetVersion.ps1 -ProjectDirectory . -OutputFile .\include\Version.h
 
 [env:az-delivery-devkit-v4-OTA-0AC]
 upload_port = 192.168.178.101
@@ -26,7 +22,7 @@ upload_protocol = espota
 extra_scripts = my_extra_script.py
 build_flags = 
 	${env.build_flags}
-;lib_deps = https://github.com/scholzol/MySQL_MariaDB_Generic.git
+lib_deps = bblanchon/ArduinoJson@^7.0.4
 
 [env:az-delivery-devkit-v4-OTA-464]
 upload_port = 192.168.178.111
@@ -34,15 +30,16 @@ upload_protocol = espota
 extra_scripts = my_extra_script.py
 build_flags = 
 	${env.build_flags}
-;lib_deps = https://github.com/scholzol/MySQL_MariaDB_Generic.git
+lib_deps = bblanchon/ArduinoJson@^7.0.4
 
 [env:az-delivery-devkit-v4-OTA-8CC]
 upload_port = 192.168.178.157
 upload_protocol = espota
+monitor_port = COM11
 extra_scripts = my_extra_script.py
 build_flags = 
 	${env.build_flags}
-;lib_deps = https://github.com/scholzol/MySQL_MariaDB_Generic.git
+lib_deps = bblanchon/ArduinoJson@^7.0.4
 
 [env:az-delivery-devkit-v4-OTA-018]
 upload_port = 192.168.178.117
@@ -50,7 +47,7 @@ upload_protocol = espota
 extra_scripts = my_extra_script.py
 build_flags = 
 	${env.build_flags}
-;lib_deps = https://github.com/scholzol/MySQL_MariaDB_Generic.git
+lib_deps = bblanchon/ArduinoJson@^7.0.4
 
 [env:az-delivery-devkit-v4-OTA-6F0]
 upload_port = 192.168.178.118
@@ -58,12 +55,12 @@ upload_protocol = espota
 extra_scripts = pre:my_extra_script.py
 build_flags = 
 	${env.build_flags}
-;lib_deps = https://github.com/scholzol/MySQL_MariaDB_Generic.git
+lib_deps = bblanchon/ArduinoJson@^7.0.4
 
 [env:az-delivery-devkit-v4-Serial]
+upload_port = COM11
 upload_protocol = esptool
 extra_scripts = my_extra_script.py
 build_flags = 
 	${env.build_flags}
-;lib_deps = https://github.com/scholzol/MySQL_MariaDB_Generic.git
-
+lib_deps = bblanchon/ArduinoJson@^7.0.4
