+++ src/MySQL_1.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#ifndef MYSQL_1_H
+#define MYSQL_1_H
+
+void updateBoardTable(char ssid32[13]);
+
+#endif /* MYSQL_1_H */
