+++ src/StartServer.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#ifndef STARTSERVER
+#define STARTSERVER
+
+void startServer();
+
+#endif /* STARTSERVER */
